Code/requirements.txt
pyyaml>=6.0.1
tqdm>=4.66.0
numpy>=1.23.0
regex>=2023.10.3
